cutover to new version of walker package ranger
